requirement.txt
numpy>=1.21.0
opencv-python>=4.8.0
Pillow>=9.0.0
tqdm>=4.64.0
torch>=2.0.0
torchvision>=0.15.0
matplotlib>=3.5.0
tensorboard>=2.10.0